Fixed saving Bit values AS3: Automatic Calculating Body parameters

+++ trunk/src/com/jpexs/asdec/types/MATRIX.java
@@ -51,11 +51,11 @@ public class MATRIX {
     /**
      * X translate value in twips
      */
-    public long translateX;
+    public int translateX;
     /**
      * Y translate value in twips
      */
-    public long translateY;
+    public int translateY;
 
     /**
      * Nbits used for store translate values
